7. Size

+++ 0x00-hello_world/6-size.c
@@ -16,6 +16,6 @@ int main(void)
 	printf("Size of an int: %lu byte(s)\n", (unsigned long)Sizeof(a));
 	printf("Size of a long int: %lu byte(s)\n", (unsigned long)Sizeof(b));
 	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)Sizeof(c));
-	printf("Size of a float: %lu byte(s)\n", (int) Sizeof(f));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)Sizeof(f));
 	return (0);
 }
